+++ source/simulation2/components/CCmpRangeManager.cpp
@@ -452,14 +452,14 @@ public:
 		{
 			for (size_t i = 0; i < oldPlayerCounts.size(); ++i)
 			{
-				debug_printf(L"%d: ", i);
+				debug_printf(L"%d: ", (int)i);
 				for (size_t j = 0; j < oldPlayerCounts[i].size(); ++j)
 					debug_printf(L"%d ", oldPlayerCounts[i][j]);
 				debug_printf(L"\n");
 			}
 			for (size_t i = 0; i < m_LosPlayerCounts.size(); ++i)
 			{
-				debug_printf(L"%d: ", i);
+				debug_printf(L"%d: ", (int)i);
 				for (size_t j = 0; j < m_LosPlayerCounts[i].size(); ++j)
 					debug_printf(L"%d ", m_LosPlayerCounts[i][j]);
 				debug_printf(L"\n");
